Move MessageCellView to its own file
